feat(grid): new grid (#663)
* feat(grid): proof of concept

* feat(grid): updated css

* fix(grid): splitted css for ship/demo

* fix(grid): swapped default state

* feat(grid): removed container width from shipping

* fix(grid): renaming old grid

* docs(grid): depreaction story

* fix(grid): removed redundant styling

* fix(grid): splitted fixed and fluid stories

* fix(grid): comments for media queries

* fix(grid): updated css property to use new standard

* feat(grid): proof of concept

* feat(grid): updated css

* fix(grid): splitted css for ship/demo

* fix(grid): swapped default state

* feat(grid): removed container width from shipping

* fix(grid): renaming old grid

* docs(grid): depreaction story

* fix(grid): remove duplicate import

---------
