+++ packages/core/src/global/grid.scss
@@ -1,0 +1,74 @@
+// mobile first styling
+.tds-grid-container {
+  display: flex;
+  justify-content: center;
+}
+
+.tds-grid-fixed,
+.tds-grid-fluid {
+  display: grid;
+  width: calc(100% - 32px);
+  padding: 0 16px;
+  grid-template-columns: repeat(6, 1fr);
+  gap: 16px;
+}
+
+// mobile
+@media screen and (min-width: 416px) {
+  .tds-grid-fixed {
+    width: calc(416px - 32px);
+  }
+
+  .tds-grid-fluid {
+    grid-template-columns: repeat(12, 1fr);
+  }
+}
+
+// tablet-sm
+@media screen and (min-width: 608px) {
+  .tds-grid-fixed {
+    width: calc(608px - 32px);
+    grid-template-columns: repeat(12, 1fr);
+  }
+}
+
+// tablet-md
+@media screen and (min-width: 796px) {
+  .tds-grid-fixed {
+    width: calc(796px - 32px);
+  }
+
+  .tds-grid-fluid {
+    width: calc(100% - 64px);
+    padding: 0 32px;
+  }
+}
+
+// tablet-lg
+@media screen and (min-width: 976px) {
+  .tds-grid-fixed {
+    width: calc(976px - 32px);
+  }
+}
+
+// desktop-sm
+@media screen and (min-width: 1200px) {
+  .tds-grid-fixed {
+    width: calc(1200px - 64px);
+    padding: 0 32px;
+  }
+}
+
+// desktop-md
+@media screen and (min-width: 1440px) {
+  .tds-grid-fixed {
+    width: calc(1440px - 64px);
+  }
+}
+
+// desktop-lg
+@media screen and (min-width: 1920px) {
+  .tds-grid-fixed {
+    width: calc(1920px - 64px);
+  }
+}
